refactor: extract generated rust templates
