+++ scripts/templates/generated_mod.rs
@@ -1,0 +1,3 @@
+// @generated by scripts/update-registry.py. Do NOT modify manually.
+
+pub(crate) mod named;
