+++ scripts/update-registry.py
@@ -17,6 +17,8 @@ ASSET_CHAINS_PATH = ROOT / "assets" / "chains.json"
 GENERATED_MOD_PATH = ROOT / "src" / "generated" / "mod.rs"
 GENERATED_NAMED_PATH = ROOT / "src" / "generated" / "named.rs"
 PHF_CODEGEN_PATH = ROOT / "scripts" / "phf-codegen.rs"
+MOD_TEMPLATE_PATH = ROOT / "scripts" / "templates" / "generated_mod.rs"
+NAMED_TEMPLATE_PATH = ROOT / "scripts" / "templates" / "generated_named.rs"
 STATIC_STR_NONE = "N"
 NO_WRAPPED_NATIVE_TOKEN = "NO_WRAPPED_NATIVE_TOKEN"
 MATCHES_TAG_LIMIT = 5
@@ -204,10 +206,7 @@ def asset_chains(chains: list[Chain]) -> dict:
 
 
 def generated_mod() -> str:
-    return """// @generated by scripts/update-registry.py. Do NOT modify manually.
-
-pub(crate) mod named;
-"""
+    return MOD_TEMPLATE_PATH.read_text().format()
 
 
 def generated_named(chains: list[Chain]) -> str:
@@ -293,570 +292,34 @@ def generated_named(chains: list[Chain]) -> str:
         for tag, _flag in CHAIN_TAG_FLAGS
     }
 
-    return f"""// @generated by scripts/update-registry.py. Do NOT modify manually.
-
-use alloy_primitives::{{Address, address}};
-use alloc::string::String;
-use core::{{cmp::Ordering, fmt, str::FromStr, time::Duration}};
-
-type ChainIndex = {chain_index_type};
-
-{flag_consts}
-const {NO_WRAPPED_NATIVE_TOKEN}: u8 = u8::MAX;
-const {STATIC_STR_NONE}: StaticStr = StaticStr::NONE;
-
-#[derive(Clone, Copy)]
-#[repr(C, packed)]
-struct StaticStr {{
-    offset: u32,
-    len: u8,
-}}
-
-impl StaticStr {{
-    const NONE: Self = Self {{ offset: u32::MAX, len: 0 }};
-
-    #[inline]
-    const fn get(self) -> Option<&'static str> {{
-        if self.offset == u32::MAX {{
-            None
-        }} else {{
-            Some(self.get_unchecked())
-        }}
-    }}
-
-    #[inline]
-    const fn get_unchecked(self) -> &'static str {{
-        let ptr = unsafe {{ STRING_DATA.as_ptr().add(self.offset as usize) }};
-        let bytes = unsafe {{ core::slice::from_raw_parts(ptr, self.len as usize) }};
-        unsafe {{ core::str::from_utf8_unchecked(bytes) }}
-    }}
-}}
-
-#[derive(Clone, Copy)]
-#[repr(C, packed)]
-struct ChainData {{
-    name: StaticStr,
-    average_blocktime_millis: u16,
-    flags: {flag_type},
-    native_currency_symbol: StaticStr,
-    etherscan_api_url: StaticStr,
-    etherscan_base_url: StaticStr,
-    etherscan_api_key_name: StaticStr,
-    wrapped_native_token: u8,
-}}
-
-const fn s(offset: u32, len: u8) -> StaticStr {{
-    StaticStr {{ offset, len }}
-}}
-
-const fn d(
-    strings: [StaticStr; 5],
-    average_blocktime_millis: u16,
-    flags: {flag_type},
-    wrapped_native_token: u8,
-) -> ChainData {{
-    let [name, native_currency_symbol, etherscan_api_url, etherscan_base_url, etherscan_api_key_name] =
-        strings;
-    ChainData {{
-        name,
-        average_blocktime_millis,
-        flags,
-        native_currency_symbol,
-        etherscan_api_url,
-        etherscan_base_url,
-        etherscan_api_key_name,
-        wrapped_native_token,
-    }}
-}}
-
-const fn variant_names() -> [&'static str; {chain_data_len}] {{
-    let mut names = [""; {chain_data_len}];
-    let mut index = 0;
-    while index < CHAIN_DATA.len() {{
-        names[index] = CHAIN_DATA[index].name.get_unchecked();
-        index += 1;
-    }}
-    names
-}}
-
-/// An Ethereum EIP-155 chain.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-#[cfg_attr(feature = "schema", schemars(rename_all = "snake_case"))]
-#[repr(u64)]
-#[allow(missing_docs)]
-#[non_exhaustive]
-pub enum NamedChain {{
-{enum_variants}
-}}
-
-/// Error returned when parsing a named chain from a string fails.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub struct ParseNamedChainError;
-
-impl fmt::Display for ParseNamedChainError {{
-    #[inline]
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {{
-        f.write_str("matching variant not found")
-    }}
-}}
-
-#[cfg(feature = "std")]
-impl std::error::Error for ParseNamedChainError {{}}
-
-/// Iterator over all named chains.
-#[derive(Clone, Debug)]
-pub struct NamedChainIter {{
-    inner: core::iter::Copied<core::slice::Iter<'static, NamedChain>>,
-}}
-
-impl Default for NamedChainIter {{
-    #[inline]
-    fn default() -> Self {{
-        NamedChain::iter()
-    }}
-}}
-
-impl Iterator for NamedChainIter {{
-    type Item = NamedChain;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {{
-        self.inner.next()
-    }}
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {{
-        self.inner.size_hint()
-    }}
-}}
-
-impl DoubleEndedIterator for NamedChainIter {{
-    #[inline]
-    fn next_back(&mut self) -> Option<Self::Item> {{
-        self.inner.next_back()
-    }}
-}}
-
-impl ExactSizeIterator for NamedChainIter {{}}
-impl core::iter::FusedIterator for NamedChainIter {{}}
-
-impl NamedChain {{
-    /// The number of named chains.
-    pub const COUNT: usize = {chain_data_len};
-
-    /// All named chains in declaration order.
-    pub const VARIANTS: &'static [Self] = &[
-{variants}
-    ];
-
-    /// All named chain string representations in declaration order.
-    pub const VARIANT_NAMES: &'static [&'static str] = &VARIANT_NAMES_DATA;
-
-    /// Returns an iterator over all named chains.
-    #[inline]
-    pub fn iter() -> NamedChainIter {{
-        NamedChainIter {{ inner: Self::VARIANTS.iter().copied() }}
-    }}
-
-    #[inline]
-    const fn from_index(index: ChainIndex) -> Self {{
-        Self::VARIANTS[{chain_index_from}]
-    }}
-
-    #[inline]
-    const fn data(self) -> ChainData {{
-        match self {{
-{chain_data_arms}
-        }}
-    }}
-
-    #[inline]
-    fn from_chain_id_map(id: u64) -> Option<Self> {{
-        CHAIN_IDS.get(&id).copied().map(Self::from_index)
-    }}
-
-    #[inline]
-    const fn has_flag(self, flag: {flag_type}) -> bool {{
-        self.data().flags & flag != 0
-    }}
-
-    /// Returns the chain for the given EIP-155 chain ID.
-    #[inline]
-    pub const fn from_chain_id(id: u64) -> Option<Self> {{
-        match id {{
-{chain_id_arms}
-            _ => None,
-        }}
-    }}
-
-    /// Returns the string representation of the chain.
-    #[inline]
-    pub const fn as_str(&self) -> &'static str {{
-        self.data().name.get_unchecked()
-    }}
-
-    /// Returns `true` if this chain is Ethereum or an Ethereum testnet.
-    #[inline]
-    pub const fn is_ethereum(&self) -> bool {{
-        {tag_predicates["ethereum"]}
-    }}
-
-    /// Returns true if the chain contains Optimism configuration.
-    #[inline]
-    pub const fn is_optimism(self) -> bool {{
-        {tag_predicates["optimism"]}
-    }}
-
-    /// Returns true if the chain contains Gnosis configuration.
-    #[inline]
-    pub const fn is_gnosis(self) -> bool {{
-        {tag_predicates["gnosis"]}
-    }}
-
-    /// Returns true if the chain contains Polygon configuration.
-    #[inline]
-    pub const fn is_polygon(self) -> bool {{
-        {tag_predicates["polygon"]}
-    }}
-
-    /// Returns true if the chain contains Arbitrum configuration.
-    #[inline]
-    pub const fn is_arbitrum(self) -> bool {{
-        {tag_predicates["arbitrum"]}
-    }}
-
-    /// Returns true if the chain contains Elastic Network configuration.
-    #[inline]
-    pub const fn is_elastic(self) -> bool {{
-        {tag_predicates["elastic"]}
-    }}
-
-    /// Returns true if the chain contains Tempo configuration.
-    #[inline]
-    pub const fn is_tempo(self) -> bool {{
-        {tag_predicates["tempo"]}
-    }}
-
-    /// Returns true if the chain uses a custom Sourcify-compatible API.
-    #[inline]
-    pub const fn is_custom_sourcify(self) -> bool {{
-        {tag_predicates["custom_sourcify"]}
-    }}
-
-    /// Returns the chain's average blocktime, if applicable.
-    #[inline]
-    pub const fn average_blocktime_hint(self) -> Option<Duration> {{
-        let millis = self.data().average_blocktime_millis;
-        if millis == 0 {{
-            None
-        }} else {{
-            Some(Duration::from_millis(millis as u64))
-        }}
-    }}
-
-    /// Returns whether the chain implements EIP-1559.
-    #[inline]
-    pub const fn is_legacy(self) -> bool {{
-        self.has_flag(FLAG_LEGACY)
-    }}
-
-    /// Returns whether the chain supports the Shanghai hardfork.
-    #[inline]
-    pub const fn supports_shanghai(self) -> bool {{
-        self.has_flag(FLAG_SUPPORTS_SHANGHAI)
-    }}
-
-    /// Returns whether the chain is a testnet.
-    #[inline]
-    pub const fn is_testnet(self) -> bool {{
-        self.has_flag(FLAG_TESTNET)
-    }}
-
-    /// Returns the symbol of the chain's native currency.
-    #[inline]
-    pub const fn native_currency_symbol(self) -> Option<&'static str> {{
-        self.data().native_currency_symbol.get()
-    }}
-
-    /// Returns the chain's blockchain explorer and its API URLs.
-    #[inline]
-    pub const fn etherscan_urls(self) -> Option<(&'static str, &'static str)> {{
-        let data = self.data();
-        match (data.etherscan_api_url.get(), data.etherscan_base_url.get()) {{
-            (Some(api), Some(base)) => Some((api, base)),
-            _ => None,
-        }}
-    }}
-
-    /// Returns the chain's blockchain explorer API key environment variable name.
-    #[inline]
-    pub const fn etherscan_api_key_name(self) -> Option<&'static str> {{
-        self.data().etherscan_api_key_name.get()
-    }}
-
-    /// Returns the chain's blockchain explorer API key from the environment.
-    #[cfg(feature = "std")]
-    pub fn etherscan_api_key(self) -> Option<String> {{
-        self.etherscan_api_key_name().and_then(|name| std::env::var(name).ok())
-    }}
-
-    /// Returns the address of the public DNS node list for the given chain.
-    pub fn public_dns_network_protocol(self) -> Option<String> {{
-        const DNS_PREFIX: &str =
-            "enrtree://AKA3AM6LPBYEUDMVNU3BSVQJ5AD45Y7YPOHJLEF6W26QOE4VTUDPE@";
-        if matches!(
-            self,
-            Self::Mainnet
-                | Self::Goerli
-                | Self::Sepolia
-                | Self::Ropsten
-                | Self::Rinkeby
-                | Self::Holesky
-                | Self::Hoodi
-        ) {{
-            let mut s = String::with_capacity(DNS_PREFIX.len() + 32);
-            s.push_str(DNS_PREFIX);
-            s.push_str("all.");
-            let chain_str = self.as_ref();
-            s.push_str(chain_str);
-            let l = s.len();
-            s[l - chain_str.len()..].make_ascii_lowercase();
-            s.push_str(".ethdisco.net");
-            Some(s)
-        }} else {{
-            None
-        }}
-    }}
-
-    /// Returns the address of the most popular wrapped native token address.
-    #[inline]
-    pub const fn wrapped_native_token(self) -> Option<Address> {{
-        let index = self.data().wrapped_native_token;
-        if index == {NO_WRAPPED_NATIVE_TOKEN} {{
-            None
-        }} else {{
-            Some(WRAPPED_NATIVE_TOKENS[index as usize])
-        }}
-    }}
-
-    #[cfg(feature = "serde")]
-    fn from_serde_str(value: &str) -> Option<Self> {{
-        PARSE_NAMES
-            .get(value)
-            .or_else(|| SERDE_NAMES.get(value))
-            .copied()
-            .map(Self::from_index)
-    }}
-}}
-
-impl From<NamedChain> for &'static str {{
-    #[inline]
-    fn from(chain: NamedChain) -> Self {{
-        (&chain).into()
-    }}
-}}
-
-impl From<&NamedChain> for &'static str {{
-    #[inline]
-    fn from(chain: &NamedChain) -> Self {{
-        chain.as_str()
-    }}
-}}
-
-impl Default for NamedChain {{
-    #[inline]
-    fn default() -> Self {{
-        Self::Mainnet
-    }}
-}}
-
-macro_rules! impl_into_numeric {{
-    ($($t:ty)+) => {{$(
-        impl From<NamedChain> for $t {{
-            #[inline]
-            fn from(chain: NamedChain) -> Self {{
-                chain as $t
-            }}
-        }}
-    )+}};
-}}
-
-impl_into_numeric!(u64 i64 u128 i128);
-#[cfg(target_pointer_width = "64")]
-impl_into_numeric!(usize isize);
-
-impl num_enum::TryFromPrimitive for NamedChain {{
-    type Primitive = u64;
-    type Error = num_enum::TryFromPrimitiveError<Self>;
-
-    const NAME: &'static str = "NamedChain";
-
-    #[inline]
-    fn try_from_primitive(number: Self::Primitive) -> Result<Self, Self::Error> {{
-        Self::from_chain_id_map(number).ok_or_else(|| num_enum::TryFromPrimitiveError::new(number))
-    }}
-}}
-
-impl TryFrom<u64> for NamedChain {{
-    type Error = num_enum::TryFromPrimitiveError<Self>;
-
-    #[inline]
-    fn try_from(value: u64) -> Result<Self, Self::Error> {{
-        num_enum::TryFromPrimitive::try_from_primitive(value)
-    }}
-}}
-
-macro_rules! impl_try_from_numeric {{
-    ($($native:ty)+) => {{
-        $(
-            impl TryFrom<$native> for NamedChain {{
-                type Error = num_enum::TryFromPrimitiveError<NamedChain>;
-
-                #[inline]
-                fn try_from(value: $native) -> Result<Self, Self::Error> {{
-                    (value as u64).try_into()
-                }}
-            }}
-        )+
-    }};
-}}
-
-impl_try_from_numeric!(u8 i8 u16 i16 u32 i32 usize isize);
-
-impl fmt::Display for NamedChain {{
-    #[inline]
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {{
-        self.as_str().fmt(f)
-    }}
-}}
-
-impl AsRef<str> for NamedChain {{
-    #[inline]
-    fn as_ref(&self) -> &str {{
-        self.as_str()
-    }}
-}}
-
-impl FromStr for NamedChain {{
-    type Err = ParseNamedChainError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {{
-        PARSE_NAMES
-            .get(s)
-            .copied()
-            .map(Self::from_index)
-            .ok_or(ParseNamedChainError)
-    }}
-}}
-
-impl TryFrom<&str> for NamedChain {{
-    type Error = ParseNamedChainError;
-
-    #[inline]
-    fn try_from(value: &str) -> Result<Self, Self::Error> {{
-        value.parse()
-    }}
-}}
-
-impl PartialEq<u64> for NamedChain {{
-    #[inline]
-    fn eq(&self, other: &u64) -> bool {{
-        (*self as u64) == *other
-    }}
-}}
-
-impl PartialOrd<u64> for NamedChain {{
-    #[inline]
-    fn partial_cmp(&self, other: &u64) -> Option<Ordering> {{
-        (*self as u64).partial_cmp(other)
-    }}
-}}
-
-#[cfg(feature = "serde")]
-impl serde::Serialize for NamedChain {{
-    #[inline]
-    fn serialize<S: serde::Serializer>(&self, s: S) -> Result<S::Ok, S::Error> {{
-        s.serialize_str(self.as_ref())
-    }}
-}}
-
-#[cfg(feature = "serde")]
-impl<'de> serde::Deserialize<'de> for NamedChain {{
-    fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {{
-        struct NamedChainVisitor;
-
-        impl serde::de::Visitor<'_> for NamedChainVisitor {{
-            type Value = NamedChain;
-
-            fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {{
-                formatter.write_str("a named chain")
-            }}
-
-            fn visit_str<E: serde::de::Error>(self, value: &str) -> Result<Self::Value, E> {{
-                NamedChain::from_serde_str(value)
-                    .ok_or_else(|| serde::de::Error::unknown_variant(value, NamedChain::VARIANT_NAMES))
-            }}
-        }}
-
-        deserializer.deserialize_str(NamedChainVisitor)
-    }}
-}}
-
-#[cfg(feature = "rlp")]
-impl alloy_rlp::Encodable for NamedChain {{
-    #[inline]
-    fn encode(&self, out: &mut dyn alloy_rlp::BufMut) {{
-        (*self as u64).encode(out)
-    }}
-
-    #[inline]
-    fn length(&self) -> usize {{
-        (*self as u64).length()
-    }}
-}}
-
-#[cfg(feature = "rlp")]
-impl alloy_rlp::Decodable for NamedChain {{
-    #[inline]
-    fn decode(buf: &mut &[u8]) -> alloy_rlp::Result<Self> {{
-        let n = u64::decode(buf)?;
-        Self::try_from(n).map_err(|_| alloy_rlp::Error::Overflow)
-    }}
-}}
-
-#[cfg(feature = "arbitrary")]
-impl<'a> arbitrary::Arbitrary<'a> for NamedChain {{
-    fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {{
-        let idx = u.choose_index(Self::COUNT)?;
-        Ok(Self::VARIANTS[idx])
-    }}
-}}
-
-#[cfg(test)]
-pub(crate) const PARSE_ALIASES: &[(NamedChain, &str)] = &[
-{parse_aliases}
-];
-
-#[cfg(all(test, feature = "serde"))]
-pub(crate) const SERDE_ALIASES: &[(NamedChain, &str)] = &[
-{serde_aliases}
-];
-
-static STRING_DATA: &[u8] = {string_data};
-
-static CHAIN_DATA: [ChainData; {chain_data_len}] = [
-{chain_data}
-];
-
-static WRAPPED_NATIVE_TOKENS: [Address; {wrapped_native_token_len}] = [
-{wrapped_native_token_data}
-];
-
-static VARIANT_NAMES_DATA: [&str; {chain_data_len}] = variant_names();
-
-{phf_maps}
-"""
+    return NAMED_TEMPLATE_PATH.read_text().format(
+        chain_data=chain_data,
+        chain_data_arms=chain_data_arms,
+        chain_data_len=chain_data_len,
+        chain_id_arms=chain_id_arms,
+        chain_index_from=chain_index_from,
+        chain_index_type=chain_index_type,
+        custom_sourcify_predicate=tag_predicates["custom_sourcify"],
+        elastic_predicate=tag_predicates["elastic"],
+        enum_variants=enum_variants,
+        ethereum_predicate=tag_predicates["ethereum"],
+        flag_consts=flag_consts,
+        flag_type=flag_type,
+        gnosis_predicate=tag_predicates["gnosis"],
+        NO_WRAPPED_NATIVE_TOKEN=NO_WRAPPED_NATIVE_TOKEN,
+        optimism_predicate=tag_predicates["optimism"],
+        parse_aliases=parse_aliases,
+        phf_maps=phf_maps,
+        polygon_predicate=tag_predicates["polygon"],
+        arbitrum_predicate=tag_predicates["arbitrum"],
+        serde_aliases=serde_aliases,
+        STATIC_STR_NONE=STATIC_STR_NONE,
+        string_data=string_data,
+        tempo_predicate=tag_predicates["tempo"],
+        variants=variants,
+        wrapped_native_token_data=wrapped_native_token_data,
+        wrapped_native_token_len=wrapped_native_token_len,
+    )
 
 
 def average_blocktime_millis(chain: Chain) -> int:
